+++ frontend/oas-angular-app/src/app/callback/callback.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-callback',
+  templateUrl: './callback.component.html',
+  styleUrl: './callback.component.css'
+})
+export class CallbackComponent {
+  ngOnInit(): void {
+    // Maybe some business logic
+  }
+}