+++ frontend/oas-angular-app/src/app/forbidden/forbidden.component.ts
@@ -1,0 +1,12 @@
+import { Component, inject } from '@angular/core';
+import { OidcSecurityService } from 'angular-auth-oidc-client';
+
+
+@Component({
+  selector: 'app-forbidden',
+  templateUrl: './forbidden.component.html',
+  styleUrl: './forbidden.component.css'
+})
+export class ForbiddenComponent {
+  public isAuthenticated$ = inject(OidcSecurityService).isAuthenticated$;
+}