+++ frontend/oas-angular-app/src/app/protected/protected.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-protected',
+  templateUrl: './protected.component.html',
+  styleUrl: './protected.component.css'
+})
+export class ProtectedComponent {
+
+}