+++ frontend/oas-angular-app/src/app/unauthorized/unauthorized.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-unauthorized',
+  templateUrl: './unauthorized.component.html',
+  styleUrl: './unauthorized.component.css'
+})
+export class UnauthorizedComponent {
+
+}